+++ trackable/__init__.py
@@ -1,4 +1,12 @@
-"""Top-level package for trackable."""
+"""# trackable
+
+A minimalistic machine learning reporting tool.
+
+To start using trackable, import as,
+
+`from trackable import Report`
+
+"""
 
 __author__ = """Julian Garratt"""
 __email__ = 'trackable.py@gmail.com'