+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
-## 0.1.0 (2022-12-10)
+## 0.1.0 (2022-12-18)
 
-* First release on PyPI.
+* `add_model` to add a model to a report
+* `generate_report` to generate the report
+* Created initial docs